+++ WEA_BE/Services/IBookService.cs
@@ -20,7 +20,7 @@ public interface IBookService
     /// <param name="minPrice">Minimální cena pro filtrování</param>
     /// <param name="minPrice">Maximální cena pro filtrování</param>
     /// <returns>Seznam knih včetně celkového počtu záznamů.</returns>
-    (List<BookSimpleDto>, int totalRecords) GetBooks(string? title, string? author, string? genre, int? publicationYear, double? minRating, double? maxRating, int page, int pageSize, decimal? minPrice, decimal? maxPrice);
+    (List<BookSimpleDto>, int totalRecords) GetBooks(string? title, string? author, string? genre, int? publicationYear, double? minRating, double? maxRating, int page, int pageSize, double? minPrice, double? maxPrice);
 
     /// Vrací seznam oblíbených knih na základě zadaných filtrů a podporuje stránkování.
     /// </summary>
@@ -36,7 +36,7 @@ public interface IBookService
     /// <param name="pageSize">Počet položek na stránku (maximálně 100).</param>
     /// <returns>Seznam knih včetně celkového počtu záznamů.</returns>
 
-    (List<BookSimpleDto>, int totalRecords) GetFavouriteBooks(string? title, string? author, string? genre, int? publicationYear, double? minRating, double? maxRating, decimal? minPrice, decimal? maxPrice, int page, int pageSize, string userName);
+    (List<BookSimpleDto>, int totalRecords) GetFavouriteBooks(string? title, string? author, string? genre, int? publicationYear, double? minRating, double? maxRating, double? minPrice, double? maxPrice, int page, int pageSize, string userName);
 
     bool SetFavourite(int bookId, string userName);
